+++ lib/kammer/accounts.ex
@@ -623,7 +623,7 @@ defmodule Kammer.Accounts do
 
   ## Examples
 
-      iex> deliver_user_update_email_instructions(user, current_email, &url(~p"/users/settings/confirm-email/#{&1}"))
+      iex> deliver_user_update_email_instructions(user, current_email, &PublicLinks.confirm_url(conn, :email_change, &1))
       {:ok, %{to: ..., body: ...}}
 
   """