+++ lib/kammer_web.ex
@@ -18,7 +18,7 @@ defmodule KammerWeb do
   """
 
   @spec static_paths() :: [String.t()]
-  def static_paths, do: ~w(assets fonts images favicon.ico robots.txt manifest.json sw.js)
+  def static_paths, do: ~w(favicon.ico robots.txt)
 
   @spec router() :: Macro.t()
   def router do
@@ -28,7 +28,6 @@ defmodule KammerWeb do
       # Import common connection and controller functions to use in pipelines
       import Plug.Conn
       import Phoenix.Controller
-      import Phoenix.LiveView.Router
     end
   end
 
@@ -48,57 +47,6 @@ defmodule KammerWeb do
 
       import Plug.Conn
 
-      unquote(verified_routes())
-    end
-  end
-
-  @spec live_view() :: Macro.t()
-  def live_view do
-    quote do
-      use Phoenix.LiveView
-
-      unquote(html_helpers())
-    end
-  end
-
-  @spec live_component() :: Macro.t()
-  def live_component do
-    quote do
-      use Phoenix.LiveComponent
-
-      unquote(html_helpers())
-    end
-  end
-
-  @spec html() :: Macro.t()
-  def html do
-    quote do
-      use Phoenix.Component
-
-      # Import convenience functions from controllers
-      import Phoenix.Controller,
-        only: [get_csrf_token: 0, view_module: 1, view_template: 1]
-
-      # Include general helpers for rendering HTML
-      unquote(html_helpers())
-    end
-  end
-
-  defp html_helpers do
-    quote do
-      # Translation
-      use Gettext, backend: KammerWeb.Gettext
-
-      # HTML escaping functionality
-      import Phoenix.HTML
-      # Core UI components
-      import KammerWeb.CoreComponents
-
-      # Common modules used in templates
-      alias Phoenix.LiveView.JS
-      alias KammerWeb.Layouts
-
-      # Routes generation with the ~p sigil
       unquote(verified_routes())
     end
   end